1. 移动部分模组和开发板pinmap_impl.h位置 2. 统一GPIO口hal层接口 3. 统一pusle所有平台实现

+++ wiring/inc/wiring.h
@@ -84,8 +84,8 @@ bool timerIsEnd(uint32_t timerID, uint32_t time);
 
 uint32_t pulseIn(pin_t pin, uint16_t value, uint32_t timeout = 1000000L);
 uint32_t pulseInLong(pin_t pin, uint16_t value, uint32_t timeout = 1000000L);
-void analogWrite(uint16_t pin, uint16_t value);
-void analogWrite(uint16_t pin, uint16_t value, uint16_t pwm_frequency);
+void analogWrite(uint16_t pin, uint32_t value);
+void analogWrite(uint16_t pin, uint32_t value, uint32_t pwm_frequency);
 uint8_t analogWriteResolution(pin_t pin, uint8_t resolution);
 uint8_t analogWriteResolution(pin_t pin);
 uint32_t analogWriteMaxFrequency(pin_t pin);
